Move dependency substitution method to `Distribution` class
