+++ src/main/groovy/org/labkey/gradle/plugin/Distribution.groovy
@@ -18,7 +18,10 @@ package org.labkey.gradle.plugin
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.Task
+import org.gradle.api.UnknownDomainObjectException
+import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.Dependency
+import org.gradle.api.artifacts.DependencySubstitutions
 import org.gradle.api.artifacts.ModuleDependency
 import org.gradle.api.artifacts.ProjectDependency
 import org.gradle.api.file.DeleteSpec
@@ -95,7 +98,7 @@ class Distribution implements Plugin<Project>
 
         BuildUtils.addLabKeyDependency(project: project, config: "embedded", depProjectPath: BuildUtils.getEmbeddedProjectPath(project.gradle), depVersion: project.labkeyVersion, depProjectConfig: "embedded", transitive: false)
         TaskUtils.configureTaskIfPresent(project, 'artifactoryDeploy', { dependsOn(project.tasks.distribution) })
-
+        substituteModuleDependencies(project, "distribution")
     }
 
     private static void addTasks(Project project)
@@ -159,6 +162,40 @@ class Distribution implements Plugin<Project>
                 }
                 else if (dep instanceof ModuleDependency && !excludedModules.contains(dep.getName()))
                     project.dependencies.add("distribution", dep)
+        }
+    }
+
+    // See GH Issue 464
+    static void substituteModuleDependencies(Project project, String configName)
+    {
+        try {
+            project.configurations.named(configName) { Configuration config ->
+                config.resolutionStrategy.dependencySubstitution { DependencySubstitutions ds ->
+                    project.rootProject.subprojects {
+                        Project p ->
+                            {
+                                p.logger.debug("Considering substitution for ${p.path}.")
+                                if (BuildUtils.shouldBuildFromSource(p)) {
+                                    if (p.plugins.hasPlugin('org.labkey.build.module') ||
+                                            p.plugins.hasPlugin('org.labkey.build.fileModule') ||
+                                            p.plugins.hasPlugin('org.labkey.build.javaModule')
+                                    ) {
+                                        ds.substitute(ds.module("org.labkey.module:${p.name}"))
+                                                .using(variant(ds.project(p.path)) {
+                                                    attributes {
+                                                        attribute(FileModule.ARTIFACT_TYPE, FileModule.MODULE_ARTIFACT_TYPE)
+                                                    }
+                                                })
+
+                                        p.logger.info("Substituting org.labkey.module:${p.name} with ${p.path} module file")
+                                    }
+                                }
+                            }
+                    }
+                }
+            }
+        } catch (UnknownDomainObjectException ignore) {
+            project.logger.debug("No ${configName} configuration found for ${project.path}.")
         }
     }
 }
